modular lexer
