+++ src/lexer/position.rs
@@ -1,0 +1,23 @@
+/// Represents the position of a token in the source text.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Position {
+    pub line: usize,
+    pub column: usize,
+    pub offset: usize,
+}
+
+impl Position {
+    pub fn new() -> Self {
+        Self {
+            line: 1,
+            column: 1,
+            offset: 0,
+        }
+    }
+}
+
+impl Default for Position {
+    fn default() -> Self {
+        Self::new()
+    }
+}
